+++ library/localization/stt/__init__.py
@@ -4,7 +4,6 @@ from .base import STTProvider, Transcript, WordTimestamp
 from .deepl_stt import DeepLSTT
 from .local_gpu_whisper import LocalGPUWhisperSTT
 from .vastai_serverless import VastaiServerlessSTT
-from .vastai_whisper import VastaiWhisperSTT
 from .whisper_stt import WhisperSTT
 
 __all__ = [
@@ -14,6 +13,5 @@ __all__ = [
     "DeepLSTT",
     "LocalGPUWhisperSTT",
     "VastaiServerlessSTT",
-    "VastaiWhisperSTT",
     "WhisperSTT",
 ]